backend-cloned/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography
defusedxml==0.7.1
Django==4.1
django-allauth==0.44.0
django-better-admin-arrayfield==1.4.2
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
oauthlib==3.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
urllib3==1.26.11
